requirements.txt
colorama==0.4.6
imageio==2.37.0
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
scikit-image==0.25.2
scipy==1.15.2
tifffile==2025.2.18
tqdm==4.67.1
